update: rename default Notifications to Gallery
