+++ app/src/main/java/com/example/godutch/ui/gallery/GalleryViewModel.java
@@ -1,14 +1,14 @@
-package com.example.godutch.ui.notifications;
+package com.example.godutch.ui.gallery;
 
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.MutableLiveData;
 import androidx.lifecycle.ViewModel;
 
-public class NotificationsViewModel extends ViewModel {
+public class GalleryViewModel extends ViewModel {
 
     private MutableLiveData<String> mText;
 
-    public NotificationsViewModel() {
+    public GalleryViewModel() {
         mText = new MutableLiveData<>();
         mText.setValue("This is notifications fragment");
     }
